feat: change the langchain retriever implementation to use FGAFilter
