+++ packages/langchain-auth0-ai/langchain_auth0_ai/__init__.py
@@ -1,3 +1,3 @@
-from .FGARetriever import FGARetriever
+from .fga.FGARetriever import FGARetriever
 
 __all__ = ["FGARetriever"]
